Move packet classes to their own files
